Added "objects" to rooms.
Objects have multiple of their own Inventories, allowing for example,
items to be "on a table" and other items to be "under a table".

The "get" command has been broken by this, as we need a way to identify
which item is being taken when similar items exist in multiple objects.